+++ include/zephyr/kvss/nvs.h
