+++ samples/subsys/kvss/nvs/src/main.c
@@ -44,7 +44,7 @@
 #include <string.h>
 #include <zephyr/drivers/flash.h>
 #include <zephyr/storage/flash_map.h>
-#include <zephyr/fs/nvs.h>
+#include <zephyr/kvss/nvs.h>
 
 static struct nvs_fs fs;
