+++ samples/subsys/kvss/zms/src/main.c
@@ -13,7 +13,7 @@
 #include <string.h>
 #include <zephyr/drivers/flash.h>
 #include <zephyr/storage/flash_map.h>
-#include <zephyr/fs/zms.h>
+#include <zephyr/kvss/zms.h>
 
 static struct zms_fs fs;
