+++ subsys/settings/include/settings/settings_nvs.h
@@ -8,7 +8,7 @@
 #ifndef __SETTINGS_NVS_H_
 #define __SETTINGS_NVS_H_
 
-#include <zephyr/fs/nvs.h>
+#include <zephyr/kvss/nvs.h>
 #include <zephyr/settings/settings.h>
 
 #ifdef __cplusplus
