+++ subsys/settings/include/settings/settings_zms.h
@@ -6,7 +6,7 @@
 #ifndef __SETTINGS_ZMS_H_
 #define __SETTINGS_ZMS_H_
 
-#include <zephyr/fs/zms.h>
+#include <zephyr/kvss/zms.h>
 #include <zephyr/settings/settings.h>
 
 #ifdef __cplusplus
