+++ tests/subsys/kvss/nvs/src/main.c
@@ -9,7 +9,7 @@
 #include <zephyr/ztest.h>
 
 #include <zephyr/drivers/flash.h>
-#include <zephyr/fs/nvs.h>
+#include <zephyr/kvss/nvs.h>
 #include <zephyr/stats/stats.h>
 #include <zephyr/storage/flash_map.h>
 #include <zephyr/sys/crc.h>
