+++ tests/subsys/kvss/zms/src/main.c
@@ -10,7 +10,7 @@
 #include <zephyr/ztest.h>
 
 #include <zephyr/drivers/flash.h>
-#include <zephyr/fs/zms.h>
+#include <zephyr/kvss/zms.h>
 #include <zephyr/stats/stats.h>
 #include <zephyr/storage/flash_map.h>
 #include <zephyr/sys/crc.h>
